# build, lib/build: Build faster, auto detect program name