--- a/lib/build/build.go
+++ b/lib/build/build.go
@@ -9,6 +9,8 @@ package build
 import (
 	"fmt"
 	"log"
+	"os"
+	"path/filepath"
 	"regexp"
 	"runtime"
 	"strconv"
@@ -72,6 +74,10 @@ func setBuildData() {
 
 	stamp, _ := strconv.Atoi(Stamp)
 	Date = time.Unix(int64(stamp), 0)
+
+	if exe, err := os.Executable(); err == nil {
+		Program = strings.TrimSuffix(strings.ToLower(filepath.Base(exe)), ".exe")
+	}
 
 	date := Date.UTC().Format("2006-01-02 15:04:05 MST")
 	LongVersion = fmt.Sprintf(`%s %s "%s" (%s %s-%s) %s@%s %s`, Program, Version, Codename, runtime.Version(), runtime.GOOS, runtime.GOARCH, User, Host, date)