@AudriusButkevicius 2020-04-15
I'd rather then start hardcoding program names in each binary, then do this stuff. People can name their stuff whatever they want, last thing I want is `lol-ma-look-at-me.exe 1.4.2` showing up in random places.