@imsodin 2020-04-16
This needs some kind of notification in the changelog for packagers (e.g. @aviau, don't know who the arch etc people are ), I'll try to think of it when it's time (or maybe solve that with a way to persist changelogs other than github releases, if it's not too much work I believe that would have merit).