@calmh 2020-04-16
I don't think it does. Yes, it's no longer required to set the `Program` variable in the linker options, but there is no harm in doing so (it won't error) and everything will work as it should.