@imsodin 2020-04-16
Ah cool - I wrongly assumed that it will complain as `build.Program` doesn't exist anymore. Sorry for the noise.